requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
openai>=1.55.0
python-dotenv>=1.0.0
httpx>=0.27.0
streamlit>=1.28.0
pandas>=2.0.0
